add expenses and isolate cards functions
