+++ src/pages/dashboard/index.tsx
@@ -1,66 +1,16 @@
-import { useMemo } from "react";
-
 import { useFinanceContext } from "../../contexts/financeContext";
 import TotalCard from "../../components/totalCard";
-import {
-  endOfMonth,
-  isWithinInterval,
-  startOfMonth,
-  subMonths,
-} from "date-fns";
+
 import PrimaryCard from "../../components/primaryCard";
 import Table from "../../components/table";
+import cards from "../../utils/cards";
 
 export const Dashboard = () => {
   const { records } = useFinanceContext();
 
-  const totalPrevMonth = useMemo(() => {
-    let totalAmount = 0;
-    const currentDate = new Date();
-    const prevMonthStart = startOfMonth(subMonths(currentDate, 1));
-    const prevMonthEnd = endOfMonth(subMonths(currentDate, 1));
-
-    records.forEach((record) => {
-      if (
-        isWithinInterval(record.date, {
-          start: prevMonthStart,
-          end: prevMonthEnd,
-        })
-      ) {
-        totalAmount += record.amount;
-      }
-    });
-
-    return totalAmount;
-  }, [records]);
-
-  const totalCurrentMonth = useMemo(() => {
-    let totalAmount = 0;
-    const currentDate = new Date();
-    const currentMonthStart = startOfMonth(currentDate);
-    const currentMonthEnd = endOfMonth(currentDate);
-
-    records.forEach((record) => {
-      if (
-        isWithinInterval(record.date, {
-          start: currentMonthStart,
-          end: currentMonthEnd,
-        })
-      ) {
-        totalAmount += record.amount;
-      }
-    });
-
-    return totalAmount;
-  }, [records]);
-
-  const totalAverage = useMemo(() => {
-    let totalAmount = 0;
-    records.forEach((record) => {
-      totalAmount += record.amount;
-    });
-    return totalAmount;
-  }, [records]);
+  const { totalAverage, totalCurrentMonth, totalPrevMonth } = cards({
+    records,
+  });
 
   return (
     <div>
